Добавлены фильтры для входящих

+++ AutoWorkspaceWFA/InMessagesForm.cs
@@ -1,4 +1,7 @@
-﻿using System;
+﻿using AutoWorkplaceLib.Data;
+using AutoWorkplaceLib.Models;
+using AutoWorkplaceLib.Repositories;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
@@ -7,24 +10,23 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
-using AutoWorkplaceLib.Data;
-using AutoWorkplaceLib.Models;
-using AutoWorkplaceLib.Repositories;
 
 namespace AutoWorkspaceWFA
 {
     public partial class InMessagesForm : Form
     {
         private List<IncomingMessage> messages = new List<IncomingMessage>();
-        private readonly List<Source> sources = new List<Source>();
         readonly private InMessageRepository messageRepository;
-        private readonly string text2 = "Отправитель";
-        private readonly string text3 = "Кому";
-        private readonly string text4 = "От кого";
+        private readonly string textNumber = "Номер";
+        private readonly string textSender = "Отправитель";
+        private readonly string textRecipient = "Получатель";
+        private readonly string textAdressee = "Кому";
+        private readonly List<Source> sources = new List<Source>();
         public InMessagesForm()
         {
-            messageRepository = new InMessageRepository();
             InitializeComponent();
+            messageRepository = new InMessageRepository();
+            messages = messageRepository.SelectAll();
             dataGridView.AllowUserToResizeColumns = false;
             dataGridView.AllowUserToResizeRows = false;
             dataGridView.AllowUserToAddRows = false;
@@ -32,37 +34,53 @@ namespace AutoWorkspaceWFA
             dataGridView.AllowUserToOrderColumns = false;
             dataGridView.AutoGenerateColumns = true;
             dataGridView.ReadOnly = true;
-            makeHelpTextVisible(textBoxSender, text2);
-            makeHelpTextVisible(textBoxRecipient, text3);
-            makeHelpTextVisible(textBoxAdressee, text4);
+            groupBox1.Text = null;
+
             SourcesRepository sourcesRepository = new SourcesRepository();
             sources = sourcesRepository.AllItems.OrderBy(s => s.Id).ToList();
             //sources = new List<Source>() { new Source(1, "Электронная почта"), new Source(2, "Бумажная почта"), new Source(3, "СМДО") };
             comboBoxSource.Items.AddRange(sources.Select(s => s.Name).ToArray());
-            loadMessages();
+            loadMessages(messages);
+            dateTimePickerFrom.Value = messages.OrderBy(m => m.Date).FirstOrDefault().Date;
+            dateTimePickerTo.Value = messages.OrderByDescending(m => m.Date).FirstOrDefault().Date;
+            makeHelpTextVisible(textBoxSender, textSender);
+            makeHelpTextVisible(textBoxRecipient, textRecipient);
+            makeHelpTextVisible(textBoxAdressee, textAdressee);
+            makeHelpTextVisible(textBoxSenderFilter, textSender);
+            makeHelpTextVisible(textBoxRecipientFilter, textRecipient);
+            makeHelpTextVisible(textBoxAdresseeFilter, textAdressee);
         }
 
         private void DeleteButton_Click(object sender, EventArgs e)
         {
-            if (dataGridView.CurrentCell.RowIndex >= 0)
+            try
             {
-                messageRepository.Delete(messages[dataGridView.CurrentCell.RowIndex].Id);
-                loadMessages();
+                if (dataGridView.CurrentCell.RowIndex >= 0 && dataGridView.CurrentCell.RowIndex < messages.Count)
+                {
+                    messageRepository.Delete(messages[dataGridView.CurrentCell.RowIndex].Id);
+                    //messages = messageRepository.SelectAll();
+                    //loadMessages(messages);
+                    loadFiltered();
+                }
             }
-            loadMessages();
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+            }
         }
-        private void loadMessages()
+
+        private void loadMessages(List<IncomingMessage> messages)
         {
-            messages = messageRepository.SelectAll();
+            this.messages = messages;
             DataTable dt = new DataTable();
             dt.Columns.Add("Дата");
-            dt.Columns.Add("Отправитель");
             dt.Columns.Add("От кого");
+            dt.Columns.Add("Получатель");
             dt.Columns.Add("Кому");
             dt.Columns.Add("Источник");
             foreach (IncomingMessage message in messages)
             {
-                dt.Rows.Add(message.Date.ToShortDateString(), message.Sender, message.Adressee, message.Recipient, message.Source.ToString());
+                dt.Rows.Add(message.Date.ToShortDateString(), message.Sender, message.Recipient, message.Adressee, message.Source.ToString());
             }
             dataGridView.DataSource = dt.DefaultView;
             dataGridView.Columns[0].Width = 80;
@@ -75,8 +93,8 @@ namespace AutoWorkspaceWFA
         {
             if (textBox.Text.Length < 1)
             {
-                textBox.Text = text;
                 textBox.ForeColor = Color.Gray;
+                textBox.Text = text;
             }
         }
 
@@ -86,6 +104,80 @@ namespace AutoWorkspaceWFA
             {
                 textBox.Text = null;
                 textBox.ForeColor = Color.Black;
+            }
+        }
+
+        private void textBoxAdressee_Enter(object sender, EventArgs e)
+        {
+            makeHelpTextUnvisible(textBoxAdressee);
+        }
+
+        private void textBoxAdressee_Leave(object sender, EventArgs e)
+        {
+            makeHelpTextVisible(textBoxAdressee, textAdressee);
+        }
+
+        private void textBoxRecipient_Enter(object sender, EventArgs e)
+        {
+            makeHelpTextUnvisible(textBoxRecipient);
+        }
+
+        private void textBoxRecipient_Leave(object sender, EventArgs e)
+        {
+            makeHelpTextVisible(textBoxRecipient, textRecipient);
+        }
+
+        private void textBoxSender_Leave(object sender, EventArgs e)
+        {
+            makeHelpTextVisible(textBoxSender, textSender);
+        }
+
+        private void textBoxSender_Enter(object sender, EventArgs e)
+        {
+            makeHelpTextUnvisible(textBoxSender);
+        }
+
+        private void addButton_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                messageRepository.Insert(new IncomingMessage(
+                    0,
+                    DateTime.Now,
+                    textBoxSender.Text,
+                    textBoxRecipient.Text,
+                    textBoxAdressee.Text,
+                    comboBoxSource.SelectedIndex + 1
+                    ));
+                //messages = messageRepository.SelectAll();
+                //loadMessages(messages);
+                loadFiltered();
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+            }
+        }
+
+        private void saveButton_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                messageRepository.Update(new IncomingMessage(
+                messages[dataGridView.CurrentCell.RowIndex].Id,
+                DateTime.Now,
+                textBoxSender.Text,
+                textBoxRecipient.Text,
+                textBoxAdressee.Text,
+                comboBoxSource.SelectedIndex + 1
+                ));
+                //messages = messageRepository.SelectAll();
+                //loadMessages(messages);
+                loadFiltered();
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
             }
         }
 
@@ -106,60 +198,113 @@ namespace AutoWorkspaceWFA
             }
         }
 
-        private void textBoxAdressee_Enter(object sender, EventArgs e)
+        private void dateTimePickerFrom_ValueChanged(object sender, EventArgs e)
         {
-            makeHelpTextUnvisible(textBoxAdressee);
+            loadFiltered();
         }
 
-        private void textBoxAdressee_Leave(object sender, EventArgs e)
+        private void dateTimePickerFrom_DropDown(object sender, EventArgs e)
         {
-            makeHelpTextVisible(textBoxAdressee, text4);
+            dateTimePickerFrom.ValueChanged -= dateTimePickerFrom_ValueChanged;
         }
 
-        private void textBoxRecipient_Enter(object sender, EventArgs e)
+        private void dateTimePickerFrom_CloseUp(object sender, EventArgs e)
         {
-            makeHelpTextUnvisible(textBoxRecipient);
+            dateTimePickerFrom.ValueChanged += dateTimePickerFrom_ValueChanged;
+            dateTimePickerFrom_ValueChanged(sender, e);
         }
 
-        private void textBoxRecipient_Leave(object sender, EventArgs e)
+        private void dateTimePickerTo_ValueChanged(object sender, EventArgs e)
         {
-            makeHelpTextVisible(textBoxRecipient, text3);
+            loadFiltered();
+        }
+        private void dateTimePickerTo_DropDown(object sender, EventArgs e)
+        {
+            dateTimePickerTo.ValueChanged -= dateTimePickerTo_ValueChanged;
         }
 
-        private void textBoxSender_Leave(object sender, EventArgs e)
+        private void dateTimePickerTo_CloseUp(object sender, EventArgs e)
         {
-            makeHelpTextVisible(textBoxSender, text2);
+            dateTimePickerTo.ValueChanged += dateTimePickerTo_ValueChanged;
+            dateTimePickerTo_ValueChanged(sender, e);
         }
 
-        private void textBoxSender_Enter(object sender, EventArgs e)
+        private void textBoxSenderFilter_TextChanged(object sender, EventArgs e)
         {
-            makeHelpTextUnvisible(textBoxSender);
+            loadFiltered();
         }
 
-        private void addButton_Click(object sender, EventArgs e)
+        private void textBoxRecipientFilter_TextChanged(object sender, EventArgs e)
         {
-            messageRepository.Insert(new IncomingMessage(
-                0,
-                DateTime.Now,
-                textBoxSender.Text,
-                textBoxRecipient.Text,
-                textBoxAdressee.Text,
-                comboBoxSource.SelectedIndex + 1
-                ));
-            loadMessages();
+            loadFiltered();
         }
 
-        private void saveButton_Click(object sender, EventArgs e)
+        private void textBoxAdresseeFilter_TextChanged(object sender, EventArgs e)
         {
-            messageRepository.Update(new IncomingMessage(
-                messages[dataGridView.CurrentCell.RowIndex].Id,
-                DateTime.Now,
-                textBoxSender.Text,
-                textBoxRecipient.Text,
-                textBoxAdressee.Text,
-                comboBoxSource.SelectedIndex + 1
-                ));
-            loadMessages();
+            loadFiltered();
+        }
+
+        private void textBoxSenderFilter_Enter(object sender, EventArgs e)
+        {
+            makeHelpTextUnvisible(textBoxSenderFilter);
+        }
+
+        private void textBoxSenderFilter_Leave(object sender, EventArgs e)
+        {
+            makeHelpTextVisible(textBoxSenderFilter, textSender);
+        }
+
+        private void textBoxRecipientFilter_Enter(object sender, EventArgs e)
+        {
+            makeHelpTextUnvisible(textBoxRecipientFilter);
+        }
+
+        private void textBoxRecipientFilter_Leave(object sender, EventArgs e)
+        {
+            makeHelpTextVisible(textBoxRecipientFilter, textRecipient);
+        }
+
+        private void textBoxAdresseeFilter_Enter(object sender, EventArgs e)
+        {
+            makeHelpTextUnvisible(textBoxAdresseeFilter);
+        }
+
+        private void textBoxAdresseeFilter_Leave(object sender, EventArgs e)
+        {
+            makeHelpTextVisible(textBoxAdresseeFilter, textAdressee);
+        }
+        private void loadFiltered()
+        {
+            var messages = messageRepository.SelectAll()
+                .Where(m => m.Date > dateTimePickerFrom.Value && m.Date < dateTimePickerTo.Value);
+            if (textBoxSenderFilter.Text.Length > 0 && textBoxSenderFilter.ForeColor == Color.Black)
+            {
+                messages = messages.Where(m => m.Sender.Contains(textBoxSenderFilter.Text));
+            }
+            if (textBoxRecipientFilter.Text.Length > 0 && textBoxRecipientFilter.ForeColor == Color.Black)
+            {
+                messages = messages.Where(m => m.Sender.Contains(textBoxRecipientFilter.Text));
+            }
+            if (textBoxAdresseeFilter.Text.Length > 0 && textBoxAdresseeFilter.ForeColor == Color.Black)
+            {
+                messages = messages.Where(m => m.Adressee.Contains(textBoxAdresseeFilter.Text));
+            }
+            this.messages = messages.ToList();
+            loadMessages(this.messages);
+        }
+
+        private void resetButton_Click(object sender, EventArgs e)
+        {
+            messages = messageRepository.SelectAll().ToList();
+            dateTimePickerFrom.Value = messages.OrderBy(m => m.Date).FirstOrDefault().Date;
+            dateTimePickerTo.Value = messages.OrderByDescending(m => m.Date).FirstOrDefault().Date;
+            textBoxSenderFilter.Text = null;
+            textBoxRecipientFilter.Text = null;
+            textBoxAdresseeFilter.Text = null;
+            makeHelpTextVisible(textBoxSenderFilter, textSender);
+            makeHelpTextVisible(textBoxRecipientFilter, textRecipient);
+            makeHelpTextVisible(textBoxAdresseeFilter, textAdressee);
+            loadMessages(messages);
         }
     }
 }
